requirements.txt
typer==0.9.0
rich==13.5.2
loguru==0.7.2
openai==1.30.1
langchain==0.2.0
chromadb>=0.4.0
neo4j==5.17.0
python-dotenv==1.0.1
msal==1.25.0
requests==2.31.0
anthropic>=0.25.0